✨ feat: 完成仪表盘字段补全以及大部分规则静态页面

+++ src/app.vue
@@ -32,20 +32,14 @@
     const language = computed(() => appStore.language);
     const locale = computed(() => (language.value === 'zh-CN' ? zhCn : en));
 
-    const versionStorage = useLocalStorage('version', version);
-
-    if (versionStorage.value !== version) {
-        versionStorage.value = version;
-        // 重置本地存储
-        localStorage.clear();
-        // 清空所有cookie
-        let cookies = document.cookie.split(';');
-
-        for (let i = 0; i < cookies.length; i++) {
-            let cookie = cookies[i];
-            let eqPos = cookie.indexOf('=');
-            let name = eqPos > -1 ? cookie.substr(0, eqPos) : cookie;
-            document.cookie = name + '=;expires=Thu, 01 Jan 1970 00:00:00 GMT;path=/';
+    const versionStorage = useCookie('version', {
+        default: () => version,
+    });
+    if (process.client) {
+        if (versionStorage.value !== version) {
+            versionStorage.value = version;
+            // 重置本地存储
+            localStorage.clear();
         }
     }
